==> experiment/00-data.ipynb <==
from pathlib import Path
import os

# create dir processed_data if it doesn't exist
processed_data_dir = Path("processed_data")
if not processed_data_dir.exists():
    processed_data_dir.mkdir()

import pandas as pd
import numpy as np

from sklearn.preprocessing import LabelEncoder

df = pd.read_csv('datasets/Bala_classification_dataset.csv')
df = df.drop(labels=["Me3"], axis=1)
df_category = df.select_dtypes(exclude=[np.number])

le = LabelEncoder()
for col in df_category.columns:
    df[col] = le.fit_transform(df[col])
df.to_csv('processed_data/Bala_classification_dataset.csv', index=False)
df

from sklearn.preprocessing import LabelEncoder

df = pd.read_csv('datasets/Bala_regression_dataset.csv')
df_category = df.select_dtypes(exclude=[np.number])

le = LabelEncoder()
for col in df_category.columns:
    df[col] = le.fit_transform(df[col])
df.to_csv('processed_data/Bala_regression_dataset.csv', index=False)
df

from sklearn.preprocessing import LabelEncoder

df = pd.read_csv('datasets/bandgap.csv')
df_category = df.select_dtypes(exclude=[np.number])

le = LabelEncoder()
for col in df_category.columns:
    df[col] = le.fit_transform(df[col])
df.to_csv('processed_data/bandgap.csv', index=False)
df

from sklearn.preprocessing import LabelEncoder

df = pd.read_csv('datasets/BMDS_data.csv')
df = df.drop(['materials', 'bilayer_materials', 'ground_state_Stacking', 'space_group'], axis=1)
df_category = df.select_dtypes(exclude=[np.number])

le = LabelEncoder()
for col in df_category.columns:
    df[col] = le.fit_transform(df[col])
df.to_csv('processed_data/BMDS_data.csv', index=False)
df

from sklearn.preprocessing import LabelEncoder

df = pd.read_csv('datasets/Crystal_structure.csv')
df = df.drop(columns=['Compound', 'A', 'B', 'In literature'])
df_category = df.select_dtypes(exclude=[np.number])

le = LabelEncoder()
for col in df_category.columns:
    df[col] = le.fit_transform(df[col])
df.to_csv('processed_data/Crystal_structure.csv', index=False)
df

from sklearn.preprocessing import LabelEncoder

df = pd.read_csv('datasets/Glass.csv')

df["Type of glass"] = df["Type of glass"].astype("category")

df_category = df.select_dtypes(exclude=[np.number])

le = LabelEncoder()
for col in df_category.columns:
    df[col] = le.fit_transform(df[col])
df.to_csv('processed_data/Glass.csv', index=False)
df

from sklearn.preprocessing import LabelEncoder

df = pd.read_csv('datasets/PUE.csv')
df = df.drop(['SSID'], axis=1)
df_category = df.select_dtypes(exclude=[np.number])

le = LabelEncoder()
for col in df_category.columns:
    df[col] = le.fit_transform(df[col])
df.to_csv('processed_data/PUE.csv', index=False)
df
