==> experiment/04-03-Shap-analysis.ipynb <==
# read from pickle file
import pickle

results = {}

with open("results/dcc_fi_results-p.pkl", "rb") as f:
    results = pickle.load(f)

results

# sort results["PUE.csv"]["Deletion"]['feats2shap']

sorted(results["PUE.csv"]['ori_feats2shap'].items(), key=lambda x: x[1], reverse=True)

sorted(results["PUE.csv"]["Deletion"]['feats2shap'].items(), key=lambda x: x[1], reverse=True)

diff = {}
for k, v in results["PUE.csv"]['ori_feats2shap'].items():
    diff[k] = abs(v - results["PUE.csv"]["Deletion"]['feats2shap'][k])

sorted_diff = sorted(diff.items(), key=lambda x: x[1], reverse=True)
sorted_diff

sorted_dcc = sorted(results["PUE.csv"]["Deletion"]['dcc'].items(), key=lambda x: x[1], reverse=True)
sorted_dcc

import pandas as pd

df = pd.read_csv("processed_data/PUE.csv")
df.corr()["logYM"].abs().sort_values(ascending=False)
