==> experiment/07-external-dataset-process.ipynb <==
import pandas as pd
# 数据集来源 https://github.com/cakshat/AlloyBERT/blob/main/data/MPEA_numeric/MPEA_numeric.csv
# https://doi.org/10.1038/s41597-020-00768-9
df = pd.read_csv("datasets/MPEA_numeric.csv")
df.drop(columns=["PROPERTY: Calculated Young modulus (GPa)","PROPERTY: Calculated Density (g/cm$^3$)"], inplace=True)


from sklearn.model_selection import KFold
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import r2_score
import numpy as np

X = df.drop(columns=["PROPERTY: YS (MPa)"])
y = df["PROPERTY: YS (MPa)"]
kf = KFold(n_splits=5, shuffle=True, random_state=0)
rf = RandomForestRegressor(n_estimators=1000, random_state=0)
scrores = []
for train_index, test_index in kf.split(X):
    X_train, X_test = X.iloc[train_index], X.iloc[test_index]
    y_train, y_test = y.iloc[train_index], y.iloc[test_index]
    rf.fit(X_train, y_train)
    y_pred = rf.predict(X_test)
    scrores.append(r2_score(y_test, y_pred))

print(f"R2 scores: {np.mean(scrores):.3f} ± {np.std(scrores):.3f}")

df.to_csv("external-datasets/MPEA.csv", index=False)
