==> experiment/09-external.ipynb <==
import pandas as pd

df = pd.read_csv("external/Case 1.csv")
df

from sklearn.preprocessing import LabelEncoder
le = LabelEncoder()
le.fit(df['Formability'])
df['Formability'] = le.transform(df['Formability'])
df.to_csv("Case_1_process.csv", index=False)

from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split, KFold

model = RandomForestClassifier()
X = df.drop(columns=['Formability'])
y = df['Formability']
kf = KFold(n_splits=5, shuffle=True, random_state=42)
scores = []
for train_index, test_index in kf.split(X):
    X_train, X_test = X.iloc[train_index], X.iloc[test_index]
    y_train, y_test = y.iloc[train_index], y.iloc[test_index]
    model.fit(X_train, y_train)
    score = model.score(X_test, y_test)
    scores.append(score)
print("Cross-validation scores:", scores)
print("Mean cross-validation score:", sum(scores)/len(scores))

import pandas as pd

df = pd.read_csv("external/RAFM-dataset.csv")
df

from sklearn.ensemble import RandomForestRegressor
from sklearn.model_selection import train_test_split, KFold

model = RandomForestRegressor()
X = df.drop(columns=['UTS'])
y = df['UTS']
kf = KFold(n_splits=5, shuffle=True, random_state=42)
scores = []
for train_index, test_index in kf.split(X):
    X_train, X_test = X.iloc[train_index], X.iloc[test_index]
    y_train, y_test = y.iloc[train_index], y.iloc[test_index]
    model.fit(X_train, y_train)
    score = model.score(X_test, y_test)
    scores.append(score)
print("Cross-validation scores:", scores)
print("Mean cross-validation score:", sum(scores)/len(scores))

import pandas as pd

df = pd.read_csv("external/RAFM-dataset-2.csv")
df

from sklearn.ensemble import RandomForestRegressor
from sklearn.model_selection import train_test_split, KFold

model = RandomForestRegressor()
X = df.drop(columns=['TE'])
y = df['TE']
kf = KFold(n_splits=5, shuffle=True, random_state=0)
scores = []
for train_index, test_index in kf.split(X):
    X_train, X_test = X.iloc[train_index], X.iloc[test_index]
    y_train, y_test = y.iloc[train_index], y.iloc[test_index]
    model.fit(X_train, y_train)
    score = model.score(X_test, y_test)
    scores.append(score)
print("Cross-validation scores:", scores)
print("Mean cross-validation score:", sum(scores)/len(scores))

from rdkit import Chem
from rdkit.Chem import AllChem, rdFingerprintGenerator, Descriptors
from rdkit.ML.Descriptors import MoleculeDescriptors
import numpy as np

def ecfp4_fingerprint(smiles):
    try:
        mol = Chem.MolFromSmiles(smiles)
        mol = Chem.MolFromSmiles(smiles)
        morgan = rdFingerprintGenerator.GetMorganGenerator(radius=2, fpSize=2048)
        fp = morgan.GetFingerprint(mol)
        arr = []
        for i in range(fp.GetNumBits()):
            arr.append(fp.GetBit(i))
        return arr
    except:
        return [np.nan]*2048

def rdkit_descriptor(smiles):
    try:
        mol = Chem.MolFromSmiles(smiles)
        mol = Chem.AddHs(mol)
        calc = MoleculeDescriptors.MolecularDescriptorCalculator([desc[0] for desc in Descriptors._descList])
        desc_values = calc.CalcDescriptors(mol)
        return desc_values
    except:
        return [np.nan]*len(Descriptors._descList)

df = pd.read_csv("external/SAMPL.csv")
df.drop(['iupac','calc'], axis=1, inplace=True)
df

fingerprints = df['smiles'].apply(ecfp4_fingerprint)
cols = [f'ECFP4_{i}' for i in range(2048)]
fp_df = pd.DataFrame(fingerprints.tolist(), columns=cols, dtype=int)

df = pd.concat([df.reset_index(drop=True), fp_df.reset_index(drop=True)], axis=1)
df.drop(columns=['smiles'], inplace=True)
df.dropna(inplace=True, axis=1)
df

from sklearn.ensemble import RandomForestRegressor
from sklearn.model_selection import train_test_split, KFold

model = RandomForestRegressor()
X = df.drop(columns=['expt'])
y = df['expt']

kf = KFold(n_splits=5, shuffle=True, random_state=42)
scores = []
for train_index, test_index in kf.split(X):
    X_train, X_test = X.iloc[train_index], X.iloc[test_index]
    y_train, y_test = y.iloc[train_index], y.iloc[test_index]
    model.fit(X_train, y_train)
    score = model.score(X_test, y_test)
    scores.append(score)
print("Cross-validation scores:", scores)
print("Mean cross-validation score:", sum(scores)/len(scores))

from sklearn.decomposition import PCA
pca = PCA(n_components=30)
fingerprint_data = df.drop(columns=['expt'])
pca_result = pca.fit_transform(fingerprint_data)
pca_df = pd.DataFrame(pca_result, columns=[f'PC{i+1}' for i in range(pca_result.shape[1])])
df = pd.concat([df.reset_index(drop=True), pca_df.reset_index(drop=True)], axis=1)
df.drop(columns=fingerprint_data.columns, inplace=True)
# df.to_csv("SAMPL_PCA.csv", index=False)

df.to_csv("external/FreeSol_process.csv", index=False)

from sklearn.ensemble import RandomForestRegressor
from sklearn.model_selection import train_test_split, KFold
model = RandomForestRegressor()
X = df.drop(columns=['expt'])
y = df['expt']

kf = KFold(n_splits=5, shuffle=True, random_state=42)
scores = []
for train_index, test_index in kf.split(X):
    X_train, X_test = X.iloc[train_index], X.iloc[test_index]
    y_train, y_test = y.iloc[train_index], y.iloc[test_index]
    model.fit(X_train, y_train)
    score = model.score(X_test, y_test)
    scores.append(score)
print("Cross-validation scores:", scores)
print("Mean cross-validation score:", sum(scores)/len(scores))

df = pd.read_csv("external/FINALTABLE.csv")
df

from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split, KFold
model = RandomForestClassifier()
X = df.drop(columns=['label'])
y = df['label']

kf = KFold(n_splits=5, shuffle=True, random_state=0)
scores = []
for train_index, test_index in kf.split(X):
    X_train, X_test = X.iloc[train_index], X.iloc[test_index]
    y_train, y_test = y.iloc[train_index], y.iloc[test_index]
    model.fit(X_train, y_train)
    score = model.score(X_test, y_test)
    scores.append(score)
print("Cross-validation scores:", scores)
print("Mean cross-validation score:", sum(scores)/len(scores))

df = pd.read_csv("external/BBBP.csv")
df.drop(['num','name'], axis=1, inplace=True)
df

fingerprints = df['smiles'].apply(ecfp4_fingerprint)
cols = [f'ECFP4_{i}' for i in range(2048)]
fp_df = pd.DataFrame(fingerprints.tolist(), columns=cols)

df = pd.concat([df.reset_index(drop=True), fp_df.reset_index(drop=True)], axis=1)
df.drop(columns=['smiles'], inplace=True)
df.dropna(inplace=True, axis=0)
# set all to int
df = df.astype(dtype=int)
df

from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split, KFold
model = RandomForestClassifier()
X = df.drop(columns=['p_np'])
y = df['p_np']

kf = KFold(n_splits=5, shuffle=True, random_state=0)
scores = []
for train_index, test_index in kf.split(X):
    X_train, X_test = X.iloc[train_index], X.iloc[test_index]
    y_train, y_test = y.iloc[train_index], y.iloc[test_index]
    model.fit(X_train, y_train)
    score = model.score(X_test, y_test)
    scores.append(score)
print("Cross-validation scores:", scores)
print("Mean cross-validation score:", sum(scores)/len(scores))

from sklearn.decomposition import PCA
pca = PCA(n_components=20)
fingerprint_data = df.drop(columns=['p_np'])
pca_result = pca.fit_transform(fingerprint_data)
pca_df = pd.DataFrame(pca_result, columns=[f'PC{i+1}' for i in range(pca_result.shape[1])])
df = pd.concat([df.reset_index(drop=True), pca_df.reset_index(drop=True)], axis=1)
df.drop(columns=fingerprint_data.columns, inplace=True)
df.to_csv("external/BBBP_PCA.csv", index=False)

from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split, KFold
model = RandomForestClassifier()
X = df.drop(columns=['p_np'])
y = df['p_np']

kf = KFold(n_splits=5, shuffle=True, random_state=0)
scores = []
for train_index, test_index in kf.split(X):
    X_train, X_test = X.iloc[train_index], X.iloc[test_index]
    y_train, y_test = y.iloc[train_index], y.iloc[test_index]
    model.fit(X_train, y_train)
    score = model.score(X_test, y_test)
    scores.append(score)
print("Cross-validation scores:", scores)
print("Mean cross-validation score:", sum(scores)/len(scores))

df = pd.read_csv("external/clintox.csv")
df

fingerprints = df['smiles'].apply(ecfp4_fingerprint)
cols = [f'ECFP4_{i}' for i in range(2048)]
fp_df = pd.DataFrame(fingerprints.tolist(), columns=cols)

df = pd.concat([df.reset_index(drop=True), fp_df.reset_index(drop=True)], axis=1)
df.drop(columns=['smiles'], inplace=True)
df.dropna(inplace=True, axis=0)
# set all to int
df = df.astype(dtype=int)
df

from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split, KFold
model = RandomForestClassifier()
X = df.drop(columns=['CT_TOX'])
y = df['CT_TOX']

kf = KFold(n_splits=5, shuffle=True, random_state=0)
scores = []
for train_index, test_index in kf.split(X):
    X_train, X_test = X.iloc[train_index], X.iloc[test_index]
    y_train, y_test = y.iloc[train_index], y.iloc[test_index]
    model.fit(X_train, y_train)
    score = model.score(X_test, y_test)
    scores.append(score)
print("Cross-validation scores:", scores)
print("Mean cross-validation score:", sum(scores)/len(scores))

from sklearn.decomposition import PCA
pca = PCA(n_components=10)
fingerprint_data = df.drop(columns=['CT_TOX'])
pca_result = pca.fit_transform(fingerprint_data)
pca_df = pd.DataFrame(pca_result, columns=[f'PC{i+1}' for i in range(pca_result.shape[1])])
df = pd.concat([df.reset_index(drop=True), pca_df.reset_index(drop=True)], axis=1)
df.drop(columns=fingerprint_data.columns, inplace=True)
df.to_csv("external/clintox_PCA.csv", index=False)

from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split, KFold
model = RandomForestClassifier()
X = df.drop(columns=['CT_TOX'])
y = df['CT_TOX']

kf = KFold(n_splits=5, shuffle=True, random_state=0)
scores = []
for train_index, test_index in kf.split(X):
    X_train, X_test = X.iloc[train_index], X.iloc[test_index]
    y_train, y_test = y.iloc[train_index], y.iloc[test_index]
    model.fit(X_train, y_train)
    score = model.score(X_test, y_test)
    scores.append(score)
print("Cross-validation scores:", scores)
print("Mean cross-validation score:", sum(scores)/len(scores))
